+++ src/api-legacy/v1/borehole/patch.py
@@ -174,6 +174,9 @@ class PatchBorehole(Action):
         elif field == 'custom.qt_depth':
             column = 'qt_depth_id_cli'
 
+        elif field == 'depth_precision':
+            column = 'qt_depth_id_cli'
+
         elif field == 'custom.qt_top_bedrock':
             column = 'qt_top_bedrock'
 
@@ -222,6 +225,7 @@ class PatchBorehole(Action):
                 'country_bho',
                 'canton_bho',
                 'municipality_bho',
+                'depth_precision',
                 'drill_diameter',
                 'custom.drill_diameter',
                 'inclination',
@@ -327,7 +331,7 @@ class PatchBorehole(Action):
                     schema = 'custom.chronostratigraphy_top_bedrock'
 
                 elif field == 'custom.qt_top_bedrock_tvd':
-                    schema = 'custom.qt_top_bedrock'
+                    schema = 'depth_precision'
 
                 elif field == 'qt_reference_elevation':
                     schema = 'elevation_precision'
@@ -336,10 +340,10 @@ class PatchBorehole(Action):
                     schema = 'reference_elevation_type'
 
                 elif field == 'qt_total_depth_tvd':
-                    schema = 'custom.qt_top_bedrock'
+                    schema = 'depth_precision'
                 
                 elif field == 'custom.qt_depth':
-                    schema = 'custom.qt_top_bedrock'
+                    schema = 'depth_precision'
 
                 # Check if domain is extracted from the correct schema
                 if value is not None and schema != (